stable version.

+++ src/main/java/io/rwwwx/footballmanagersystem/service/TeamService.java
@@ -1,12 +1,13 @@
 package io.rwwwx.footballmanagersystem.service;
 
 import io.rwwwx.footballmanagersystem.entity.Team;
-import io.rwwwx.footballmanagersystem.entity.TeamDTO;
+import io.rwwwx.footballmanagersystem.dto.TeamDTO;
+import io.rwwwx.footballmanagersystem.exception.InvalidIdException;
 import io.rwwwx.footballmanagersystem.repository.TeamRepository;
 import io.rwwwx.footballmanagersystem.utils.Mapper;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.transaction.Transactional;
 
 @Service
 @Transactional
@@ -20,20 +21,34 @@ public class TeamService {
         this.mapper = mapper;
     }
 
-    public TeamDTO saveNewTeam(TeamDTO teamDTO) {
+    public TeamDTO save(TeamDTO teamDTO) {
         return mapper.convertToDto(teamRepository.save(mapper.convertToEntity(teamDTO)));
     }
 
+    @Transactional(readOnly = true)
     public TeamDTO getById(Long id) {
         return mapper.convertToDto(teamRepository.getById(id));
     }
 
+    @Transactional
     public TeamDTO updateTeam(Long id, TeamDTO updatedTeam) {
-        return mapper.convertToDto(teamRepository.save(mapper.convertToEntity(updatedTeam)));
+        if (isTeamExists(id)) {
+            return mapper.convertToDto(teamRepository.save(mapper.convertToEntity(updatedTeam)));
+        } else {
+            throw new InvalidIdException();
+        }
     }
 
     public void deleteTeam(Long id) {
-        teamRepository.deleteById(id);
+        if (isTeamExists(id)) {
+            teamRepository.deleteById(id);
+        } else {
+            throw new InvalidIdException();
+        }
+    }
+
+    private boolean isTeamExists(Long id) {
+        return teamRepository.existsById(id);
     }
 
 }
